--- cmd/ethrex/launch/mod.rs
@@ -1,8 +1,0 @@
-#[cfg(not(feature = "l2"))]
-pub mod l1;
-#[cfg(feature = "l2")]
-pub mod l2;
-
-mod common;
-
-pub const DEFAULT_DATADIR: &str = "ethrex";
